fixing warnings/errors found with GCC

+++ Acid.Cam.v2.OSX/ac-util.cpp
@@ -643,7 +643,7 @@ void ac::Shuffle(int &index, cv::Mat &frame, std::vector<std::string> &filter_ar
     static auto rng = std::default_random_engine{};
     CallFilter(filter_array[index], frame);
     ++index;
-    if(index > filter_array.size()-1) {
+    if(static_cast<unsigned int>(index) > filter_array.size()-1) {
         index = 0;
         std::shuffle(filter_array.begin(), filter_array.end(),rng);
     }
